+++ sdk/resourcemanager/compute/armcompute/constants.go
@@ -10,15 +10,16 @@ package armcompute
 
 const (
 	moduleName    = "armcompute"
-	moduleVersion = "v5.1.0-beta.1"
+	moduleVersion = "v5.1.0"
 )
 
 type AccessLevel string
 
 const (
-	AccessLevelNone  AccessLevel = "None"
-	AccessLevelRead  AccessLevel = "Read"
-	AccessLevelWrite AccessLevel = "Write"
+	AccessLevelNone                  AccessLevel = "None"
+	AccessLevelRead                  AccessLevel = "Read"
+	AccessLevelReadForDiskInspection AccessLevel = "ReadForDiskInspection"
+	AccessLevelWrite                 AccessLevel = "Write"
 )
 
 // PossibleAccessLevelValues returns the possible values for the AccessLevel const type.
@@ -26,6 +27,7 @@ func PossibleAccessLevelValues() []AccessLevel {
 	return []AccessLevel{
 		AccessLevelNone,
 		AccessLevelRead,
+		AccessLevelReadForDiskInspection,
 		AccessLevelWrite,
 	}
 }
@@ -781,6 +783,24 @@ const (
 func PossibleExtendedLocationTypesValues() []ExtendedLocationTypes {
 	return []ExtendedLocationTypes{
 		ExtendedLocationTypesEdgeZone,
+	}
+}
+
+// FileFormat - Used to specify the file format when making request for SAS on a VHDX file format snapshot
+type FileFormat string
+
+const (
+	// FileFormatVHD - A VHD file is a disk image file in the Virtual Hard Disk file format.
+	FileFormatVHD FileFormat = "VHD"
+	// FileFormatVHDX - A VHDX file is a disk image file in the Virtual Hard Disk v2 file format.
+	FileFormatVHDX FileFormat = "VHDX"
+)
+
+// PossibleFileFormatValues returns the possible values for the FileFormat const type.
+func PossibleFileFormatValues() []FileFormat {
+	return []FileFormat{
+		FileFormatVHD,
+		FileFormatVHDX,
 	}
 }