+++ sdk/resourcemanager/compute/armcompute/disks_client.go
@@ -47,7 +47,7 @@ func NewDisksClient(subscriptionID string, credential azcore.TokenCredential, op
 // BeginCreateOrUpdate - Creates or updates a disk.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - diskName - The name of the managed disk that is being created. The name can't be changed after the disk is created. Supported
 //     characters for the name are a-z, A-Z, 0-9, _ and -. The maximum name length is 80
@@ -71,7 +71,7 @@ func (client *DisksClient) BeginCreateOrUpdate(ctx context.Context, resourceGrou
 // CreateOrUpdate - Creates or updates a disk.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 func (client *DisksClient) createOrUpdate(ctx context.Context, resourceGroupName string, diskName string, disk Disk, options *DisksClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "DisksClient.BeginCreateOrUpdate"
@@ -113,7 +113,7 @@ func (client *DisksClient) createOrUpdateCreateRequest(ctx context.Context, reso
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, disk); err != nil {
@@ -125,7 +125,7 @@ func (client *DisksClient) createOrUpdateCreateRequest(ctx context.Context, reso
 // BeginDelete - Deletes a disk.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - diskName - The name of the managed disk that is being created. The name can't be changed after the disk is created. Supported
 //     characters for the name are a-z, A-Z, 0-9, _ and -. The maximum name length is 80
@@ -147,7 +147,7 @@ func (client *DisksClient) BeginDelete(ctx context.Context, resourceGroupName st
 // Delete - Deletes a disk.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 func (client *DisksClient) deleteOperation(ctx context.Context, resourceGroupName string, diskName string, options *DisksClientBeginDeleteOptions) (*http.Response, error) {
 	var err error
 	const operationName = "DisksClient.BeginDelete"
@@ -189,7 +189,7 @@ func (client *DisksClient) deleteCreateRequest(ctx context.Context, resourceGrou
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	return req, nil
 }
@@ -197,7 +197,7 @@ func (client *DisksClient) deleteCreateRequest(ctx context.Context, resourceGrou
 // Get - Gets information about a disk.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - diskName - The name of the managed disk that is being created. The name can't be changed after the disk is created. Supported
 //     characters for the name are a-z, A-Z, 0-9, _ and -. The maximum name length is 80
@@ -245,7 +245,7 @@ func (client *DisksClient) getCreateRequest(ctx context.Context, resourceGroupNa
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -263,7 +263,7 @@ func (client *DisksClient) getHandleResponse(resp *http.Response) (DisksClientGe
 // BeginGrantAccess - Grants access to a disk.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - diskName - The name of the managed disk that is being created. The name can't be changed after the disk is created. Supported
 //     characters for the name are a-z, A-Z, 0-9, _ and -. The maximum name length is 80
@@ -288,7 +288,7 @@ func (client *DisksClient) BeginGrantAccess(ctx context.Context, resourceGroupNa
 // GrantAccess - Grants access to a disk.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 func (client *DisksClient) grantAccess(ctx context.Context, resourceGroupName string, diskName string, grantAccessData GrantAccessData, options *DisksClientBeginGrantAccessOptions) (*http.Response, error) {
 	var err error
 	const operationName = "DisksClient.BeginGrantAccess"
@@ -330,7 +330,7 @@ func (client *DisksClient) grantAccessCreateRequest(ctx context.Context, resourc
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, grantAccessData); err != nil {
@@ -339,9 +339,88 @@ func (client *DisksClient) grantAccessCreateRequest(ctx context.Context, resourc
 	return req, nil
 }
 
+// GrantAccessOnVMSSVMInstance - Grants access to a VMSS VM Instance OS disk for DiskInspection scenario.
+// If the operation fails it returns an *azcore.ResponseError type.
+//
+// Generated from API version 2023-01-02
+//   - resourceGroupName - The name of the resource group.
+//   - vmssName - The name of the VirtualmachineScaleSet managing the managed disk. The name can't be changed after the disk is
+//     created. Supported characters for the name are a-z, A-Z, 0-9, _ and -. The maximum name
+//     length is 80 characters.
+//   - diskName - The name of the managed disk that is being created. The name can't be changed after the disk is created. Supported
+//     characters for the name are a-z, A-Z, 0-9, _ and -. The maximum name length is 80
+//     characters.
+//   - grantAccessData - Access data object supplied in the body of the get disk access operation. The supported accessLevel for
+//     this action is 'ReadForDiskInspection' only.
+//   - options - DisksClientGrantAccessOnVMSSVMInstanceOptions contains the optional parameters for the DisksClient.GrantAccessOnVMSSVMInstance
+//     method.
+func (client *DisksClient) GrantAccessOnVMSSVMInstance(ctx context.Context, resourceGroupName string, vmssName string, diskName string, grantAccessData GrantAccessData, options *DisksClientGrantAccessOnVMSSVMInstanceOptions) (DisksClientGrantAccessOnVMSSVMInstanceResponse, error) {
+	var err error
+	const operationName = "DisksClient.GrantAccessOnVMSSVMInstance"
+	ctx = context.WithValue(ctx, runtime.CtxAPINameKey{}, operationName)
+	ctx, endSpan := runtime.StartSpan(ctx, operationName, client.internal.Tracer(), nil)
+	defer func() { endSpan(err) }()
+	req, err := client.grantAccessOnVMSSVMInstanceCreateRequest(ctx, resourceGroupName, vmssName, diskName, grantAccessData, options)
+	if err != nil {
+		return DisksClientGrantAccessOnVMSSVMInstanceResponse{}, err
+	}
+	httpResp, err := client.internal.Pipeline().Do(req)
+	if err != nil {
+		return DisksClientGrantAccessOnVMSSVMInstanceResponse{}, err
+	}
+	if !runtime.HasStatusCode(httpResp, http.StatusOK) {
+		err = runtime.NewResponseError(httpResp)
+		return DisksClientGrantAccessOnVMSSVMInstanceResponse{}, err
+	}
+	resp, err := client.grantAccessOnVMSSVMInstanceHandleResponse(httpResp)
+	return resp, err
+}
+
+// grantAccessOnVMSSVMInstanceCreateRequest creates the GrantAccessOnVMSSVMInstance request.
+func (client *DisksClient) grantAccessOnVMSSVMInstanceCreateRequest(ctx context.Context, resourceGroupName string, vmssName string, diskName string, grantAccessData GrantAccessData, options *DisksClientGrantAccessOnVMSSVMInstanceOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Compute/virtualMachineScaleSets/{vmssName}/disks/{diskName}/beginGetAccess"
+	if client.subscriptionID == "" {
+		return nil, errors.New("parameter client.subscriptionID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
+	if resourceGroupName == "" {
+		return nil, errors.New("parameter resourceGroupName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{resourceGroupName}", url.PathEscape(resourceGroupName))
+	if vmssName == "" {
+		return nil, errors.New("parameter vmssName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{vmssName}", url.PathEscape(vmssName))
+	if diskName == "" {
+		return nil, errors.New("parameter diskName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{diskName}", url.PathEscape(diskName))
+	req, err := runtime.NewRequest(ctx, http.MethodPost, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
+	if err != nil {
+		return nil, err
+	}
+	reqQP := req.Raw().URL.Query()
+	reqQP.Set("api-version", "2023-01-02")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	if err := runtime.MarshalAsJSON(req, grantAccessData); err != nil {
+		return nil, err
+	}
+	return req, nil
+}
+
+// grantAccessOnVMSSVMInstanceHandleResponse handles the GrantAccessOnVMSSVMInstance response.
+func (client *DisksClient) grantAccessOnVMSSVMInstanceHandleResponse(resp *http.Response) (DisksClientGrantAccessOnVMSSVMInstanceResponse, error) {
+	result := DisksClientGrantAccessOnVMSSVMInstanceResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.AccessURI); err != nil {
+		return DisksClientGrantAccessOnVMSSVMInstanceResponse{}, err
+	}
+	return result, nil
+}
+
 // NewListPager - Lists all the disks under a subscription.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - options - DisksClientListOptions contains the optional parameters for the DisksClient.NewListPager method.
 func (client *DisksClient) NewListPager(options *DisksClientListOptions) *runtime.Pager[DisksClientListResponse] {
 	return runtime.NewPager(runtime.PagingHandler[DisksClientListResponse]{
@@ -385,7 +464,7 @@ func (client *DisksClient) listCreateRequest(ctx context.Context, options *Disks
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -402,7 +481,7 @@ func (client *DisksClient) listHandleResponse(resp *http.Response) (DisksClientL
 
 // NewListByResourceGroupPager - Lists all the disks under a resource group.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - options - DisksClientListByResourceGroupOptions contains the optional parameters for the DisksClient.NewListByResourceGroupPager
 //     method.
@@ -452,7 +531,7 @@ func (client *DisksClient) listByResourceGroupCreateRequest(ctx context.Context,
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -470,7 +549,7 @@ func (client *DisksClient) listByResourceGroupHandleResponse(resp *http.Response
 // BeginRevokeAccess - Revokes access to a disk.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - diskName - The name of the managed disk that is being created. The name can't be changed after the disk is created. Supported
 //     characters for the name are a-z, A-Z, 0-9, _ and -. The maximum name length is 80
@@ -494,7 +573,7 @@ func (client *DisksClient) BeginRevokeAccess(ctx context.Context, resourceGroupN
 // RevokeAccess - Revokes access to a disk.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 func (client *DisksClient) revokeAccess(ctx context.Context, resourceGroupName string, diskName string, options *DisksClientBeginRevokeAccessOptions) (*http.Response, error) {
 	var err error
 	const operationName = "DisksClient.BeginRevokeAccess"
@@ -536,7 +615,7 @@ func (client *DisksClient) revokeAccessCreateRequest(ctx context.Context, resour
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	return req, nil
 }
@@ -544,7 +623,7 @@ func (client *DisksClient) revokeAccessCreateRequest(ctx context.Context, resour
 // BeginUpdate - Updates (patches) a disk.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - diskName - The name of the managed disk that is being created. The name can't be changed after the disk is created. Supported
 //     characters for the name are a-z, A-Z, 0-9, _ and -. The maximum name length is 80
@@ -567,7 +646,7 @@ func (client *DisksClient) BeginUpdate(ctx context.Context, resourceGroupName st
 // Update - Updates (patches) a disk.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 func (client *DisksClient) update(ctx context.Context, resourceGroupName string, diskName string, disk DiskUpdate, options *DisksClientBeginUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "DisksClient.BeginUpdate"
@@ -609,7 +688,7 @@ func (client *DisksClient) updateCreateRequest(ctx context.Context, resourceGrou
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, disk); err != nil {